refactor(audit): split extension/manifest_config structs (#5645)
